[chore] ThrowTrashView Layout Extension으로 파일 분리
